\name{ngreg-package}
\alias{ngreg-package}
\alias{ngreg}
\docType{package}
\title{
\packageTitle{ngreg}
}
\description{
\packageDescription{ngreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ngreg}
\packageIndices{ngreg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ngreg}

Maintainer: \packageMaintainer{ngreg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
